More OO.

+++ src/Aiursoft.DotDownload.Http/Services/Downloader.cs
@@ -1,62 +1,72 @@
 using System.Net.Http.Headers;
 using Aiursoft.Canon;
 using Aiursoft.DotDownload.Http.Models;
+using Aiursoft.Scanner.Abstractions;
 using Microsoft.Extensions.Logging;
 
 namespace Aiursoft.DotDownload.PluginFramework.Handlers.Download;
 
-public class Downloader
+public class Downloader : ITransientDependency
 {
+    private readonly DiskService _diskService;
     private readonly HttpClient _httpClient;
     private readonly CanonPool _downloadPool;
     private readonly CanonQueue _writePool;
     private readonly ILogger<Downloader> _logger;
 
     public Downloader(
+        DiskService diskService,
         HttpClient httpClient,
         CanonPool downloadPool,
         CanonQueue writePool,
         ILogger<Downloader> logger)
     {
+        _diskService = diskService;
         _httpClient = httpClient;
         _downloadPool = downloadPool;
         _writePool = writePool;
         _logger = logger;
     }
 
+    /// <summary>
+    /// This is a C# method that downloads a file asynchronously from a given URL and saves it to a specified location on the local disk.
+    /// </summary>
+    /// <param name="url">Http URL to download.</param>
+    /// <param name="savePath">Path to save.</param>
+    /// <param name="blockSize">Block size. It will download every blocks in parallel and save those blocks to disk with a queue.</param>
+    /// <param name="threads">Max threads to download.</param>
+    /// <param name="showProgressBar">If to show the progress bar.</param>
     public async Task DownloadAsync(
         string url,
         string savePath,
         long blockSize = 4 * 1024 * 1024,
-        int threads = 16)
+        int threads = 16,
+        bool showProgressBar = false)
     {
-        var fileToWrite = savePath;
-        if (string.IsNullOrWhiteSpace(fileToWrite))
-        {
-            fileToWrite = "./" + Path.GetFileName(url);
-        }
-
-        fileToWrite = GetAbsolutePath(Directory.GetCurrentDirectory(), fileToWrite);
 
         _logger.LogTrace($"Requesting {url}...");
         var response = await _httpClient.GetAsync(url, HttpCompletionOption.ResponseHeadersRead);
-        var contentLength = response.Content.Headers.ContentLength ?? 0;
-        _logger.LogInformation("File length: {ContentLength}MB", contentLength / 1024 / 1024);
+        var fileLength = response.Content.Headers.ContentLength ?? 0;
+        _logger.LogInformation("File length: {ContentLength}MB", fileLength / 1024 / 1024);
 
-        // TODO: What if server 301 or 302?
-        // TODO: If the file doesn't support multiple threads downloading?
-
-        // Create the file with the specified length.
-        _logger.LogInformation($"File will be saved to {fileToWrite}...");
-        File.Delete(fileToWrite);
-        await using (var fileStream = new FileStream(fileToWrite, FileMode.Create, FileAccess.Write, FileShare.None,
-                         bufferSize: 4096, useAsync: true))
+        // TODO: What if server 301 or 302? Follow redirect.
+        if (response.Headers.AcceptRanges?.Contains("bytes") != true)
         {
-            fileStream.SetLength(contentLength);
-            fileStream.Close();
+            _logger.LogWarning("The server doesn't support multiple threads downloading. Using single thread...");
+            threads = 1;
         }
 
-        var blockCount = (long)Math.Ceiling((double)contentLength / blockSize);
+        var fileToWrite = PathExtensions.GetFilePathToSave(savePath, url);
+        _logger.LogInformation($"File will be saved to {fileToWrite}...");
+
+        if (File.Exists(fileToWrite))
+        {
+            _logger.LogWarning("File already exists. Deleting...");
+            File.Delete(fileToWrite);
+        }
+        
+        _diskService.CreateFileAndAllocateSpace(fileToWrite, fileLength);
+        var blockCount = (long)Math.Ceiling((double)fileLength / blockSize);
         _logger.LogInformation("Blocks count: {BlockCount}", blockCount);
 
         var bar = new ProgressBar();
@@ -64,34 +74,34 @@ public class Downloader
         for (var i = 0; i < blockCount; i++)
         {
             var offset = i * blockSize;
-            var length = Math.Min(blockSize, contentLength - offset);
+            var length = Math.Min(blockSize, fileLength - offset);
             _downloadPool.RegisterNewTaskToPool(async () =>
             {
                 _logger.LogTrace(
-                    $"Starting download with offset: {offset / 1024 / 1024}MB, length {length / 1024 / 1024}MB, totally {contentLength / 1024 / 1024}MB");
+                    $"Starting download with offset: {offset / 1024 / 1024}MB, length {length / 1024 / 1024}MB, totally {fileLength / 1024 / 1024}MB");
                 var fileStream = await DownloadBlockAsync(url, offset, length);
                 _writePool.QueueNew(async () =>
                     {
                         _logger.LogTrace(
                             $"Saving block {offset / 1024 / 1024}MB to {(offset + length) / 1024 / 1024}MB on local disk...");
-                        await SaveBlockToDisk(fileStream, fileToWrite, offset);
+                        await _diskService.SaveBlockToDisk(fileStream, fileToWrite, offset);
                         savedBlocks++;
                         bar.Report((double)savedBlocks / blockCount);
                         _logger.LogTrace(
                             $"Finish block {offset / 1024 / 1024}MB to {(offset + length) / 1024 / 1024}MB to save on disk.");
                     },
                     startTheEngine: true,
-                    maxThreads: 1);
+                    maxThreads: 1); // Only one thread to write to disk.
                 _logger.LogTrace(
-                    $"Finished download with offset: {offset / 1024 / 1024}MB, length {length / 1024 / 1024}MB, totally {contentLength / 1024 / 1024}MB");
+                    $"Finished download with offset: {offset / 1024 / 1024}MB, length {length / 1024 / 1024}MB, totally {fileLength / 1024 / 1024}MB");
             });
         }
 
         await _downloadPool.RunAllTasksInPoolAsync(threads);
-        await _writePool.Engine;
+        await _writePool.Engine; // Make sure the write pool is finished.
         bar.Dispose();
     }
-
+    
     private async Task<MemoryStream> DownloadBlockAsync(string url, long offset, long length)
     {
         var request = new HttpRequestMessage(HttpMethod.Get, url);
@@ -103,23 +113,5 @@ public class Downloader
         await remoteStream.CopyToAsync(memoryStream);
         memoryStream.Seek(0, SeekOrigin.Begin);
         return memoryStream;
-    }
-
-    private async Task SaveBlockToDisk(MemoryStream stream, string path, long offset)
-    {
-        await using var fileStream = new FileStream(path, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.None,
-            bufferSize: 4096, useAsync: true);
-        fileStream.Seek(offset, SeekOrigin.Begin);
-        await stream.CopyToAsync(fileStream);
-        fileStream.Close();
-    }
-
-    private static string GetAbsolutePath(string currentPath, string referencePath)
-    {
-        if (Path.IsPathRooted(referencePath)) return referencePath;
-
-        referencePath = referencePath.Replace('\\', Path.DirectorySeparatorChar);
-
-        return Path.GetFullPath(Path.Combine(currentPath, referencePath));
     }
 }
